requirements.txt
# To install discord perform python3 -m pip install -U discord.py
irc3
pyyaml
asyncio-redis
